requirements.txt
torch
torchvision
omegaconf
torchmetrics
fvcore
iopath
xformers
opencv-python 
pycocotools 
matplotlib 
onnxruntime 
onnx
scipy
ffmpeg
hydra-colorlog
hydra-core
gdown
pytorch-lightning
pandas
ruamel.yaml
pyrender
wandb
distinctipy
blenderproc
einops
pin=2.6.20